refactor: Simplify SettingsViewModel initialization using stateIn
This commit refactors the `SettingsViewModel` initialization to use
`stateIn` with `SharingStarted.WhileSubscribed`. This simplifies the
code and eliminates the need for a `launch` block in the `init` block.
The `uiState` is now a `StateFlow` directly derived from the
`settingsDataStore.usernameFlow`.

Also, the test for the init function changes to accommodate the change
above.