feat: add basic login screen
This commit introduces a basic login screen that will be shown when the user is not logged in.

The `MainViewModel` now observes the login state from `UserRepository` and sets the appropriate start destination (`BookmarkListRoute` or `LoginRoute`).

The `MainActivity` is updated to use this dynamic start destination for the `NavHost`.

A new `LoginScreen.kt` and `LoginScreenViewModel.kt` have been added to handle the UI and logic for the login screen.
A new `LoginRoute` has been added to `Routes.kt`.